+++ parameters/ParameterWrapper.h
@@ -28,7 +28,7 @@ public:
 	virtual ReturnValue_t serialize(uint8_t** buffer, size_t* size,
 			const size_t max_size, bool bigEndian) const;
 
-	virtual uint32_t getSerializedSize() const;
+	virtual size_t getSerializedSize() const;
 
 	virtual ReturnValue_t deSerialize(const uint8_t** buffer, int32_t* size,
 			bool bigEndian);
